# code/backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.BoardSetup:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      setup_data:
        description: JSON string of piece positions
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: info@dotsem.be
    name: Sem Van Broekhoven
    url: https://github.com/Thomas-More-Digital-Innovation/2526-DI-004-GoStrategy
  description: This is the API server for the GoStrategy game.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  termsOfService: http://swagger.io/terms/
  title: GoStrategy API
  version: 0.1.1
paths:
...
  /users/me/board-setups:
    get:
      description: Retrieve all saved board setups for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BoardSetup'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List board setups
      tags:
      - board-setups
    post:
      consumes:
      - application/json
      description: Save a new board setup configuration
      parameters:
      - description: Board setup configuration
        in: body
        name: setup
        required: true
        schema:
          $ref: '#/definitions/models.BoardSetup'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BoardSetup'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create board setup
      tags:
      - board-setups
  /users/me/board-setups/{id}:
    delete:
      description: Delete a specific board setup configuration by ID
      parameters:
      - description: Board Setup ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete board setup
      tags:
      - board-setups
    get:
      description: Retrieve a specific board setup configuration by ID
      parameters:
      - description: Board Setup ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BoardSetup'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Board setup not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get board setup
      tags:
      - board-setups
    put:
      consumes:
      - application/json
      description: Update an existing board setup configuration by ID
      parameters:
      - description: Board Setup ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated board setup configuration
        in: body
        name: setup
        required: true
        schema:
          $ref: '#/definitions/models.BoardSetup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BoardSetup'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update board setup
      tags:
      - board-setups
...
swagger: "2.0"